requirementss.txt
absl-py==2.0.0
aiosignal==1.2.0
akro==0.0.8
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
astunparse==1.6.3
attrs==22.1.0
cached-property==1.5.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
chex==0.1.7
click==8.0.4
cloudpickle==2.2.1
cma==2.7.0
Cython==0.29.36
D4RL @ git+https://github.com/Farama-Foundation/d4rl@71a9549f2091accff93eeff68f1f3ab2c0e0a288
decorator==5.1.1
distlib==0.3.6
dm-control==1.0.14
dm-env==1.6
dm-tree==0.1.7
docker-pycreds==0.4.0
dowel==0.0.3
etils==1.3.0
fasteners==0.19
filelock==3.8.0
flatbuffers==22.9.24
flax==0.7.0
frozenlist==1.3.1
garage==2021.3.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.37
glfw==2.6.2
google-pasta==0.2.0
grpcio==1.43.0
gym==0.23.1
gym-notices==0.0.8
h5py==3.7.0
hydra-core==1.3.2
idna==3.4
imageio==2.31.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
jax==0.4.6
jax-resnet @ git+https://github.com/n2cholas/jax-resnet.git@5b00735aa0a68ec239af4a728ad4a596c1b551f6
jaxlib==0.4.6+cuda11.cudnn82
jsonschema==4.16.0
keras==2.10.0
Keras-Preprocessing==1.1.2
labmaze==1.0.6
libclang==14.0.6
lxml==4.9.3
markdown-it-py==3.0.0
mdurl==0.1.2
mjrl @ git+https://github.com/aravindr93/mjrl@3871d93763d3b49c4741e6daeaebbc605fe140dc
ml-dtypes==0.2.0
msgpack==1.0.4
mujoco==2.3.7
mujoco-py==2.1.2.14
nest-asyncio==1.5.8
networkx==2.8.6
numpy==1.24.4
omegaconf==2.3.0
opencv-python==4.8.0.76
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.1.7
packaging==23.1
pathtools==0.1.2
Pillow==10.0.1
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
protobuf==4.24.3
psutil==5.9.5
pybullet==3.2.5
pycparser==2.21
Pygments==2.16.1
PyOpenGL==3.1.7
pyparsing==3.1.1
pyrsistent==0.18.1
PyWavelets==1.4.1
PyYAML==6.0
ray==2.0.0
requests==2.31.0
rich==13.5.3
scikit-image==0.19.3
scipy==1.10.1
sentry-sdk==1.31.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.1
tabulate==0.8.10
tensorboardX==2.5.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-probability==0.18.0
tensorstore==0.1.44
termcolor==2.0.1
tifffile==2022.8.12
toolz==0.12.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.5
virtualenv==20.16.5
wandb==0.15.11
wrapt==1.14.1
zipp==3.17.0
